+++ Pierniczek/ViewModels/NewRangeDataWindowViewModel.cs
@@ -20,7 +20,7 @@ namespace Pierniczek.ViewModels
         {
         }
 
-        public decimal Min { get; set; }
-        public decimal Max { get; set; }
+        public double Min { get; set; }
+        public double Max { get; set; }
     }
 }
